third_party/blink/web_tests/http/tests/cookies/js-set-null-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Setting document.cookie to null should set a cookie named 'null'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.cookie is 'null'
PASS cookie is 'null'.
PASS successfullyParsed is true

TEST COMPLETE
